+++ lib/aspera/persistency_folder.rb
@@ -18,7 +18,8 @@ module Aspera
       Log.log.debug{"persistency folder: #{@folder}"}
     end
 
-    # @return String or nil string on existing persist, else nil
+    # Get value of persisted item
+    # @return [String,nil] Value of persisted id
     def get(object_id)
       Log.log.debug{"persistency get: #{object_id}"}
       if @cache.key?(object_id)
@@ -34,6 +35,10 @@ module Aspera
       return @cache[object_id]
     end
 
+    # Set value of persisted item
+    # @param object_id [String] Identifier of persisted item
+    # @param value [String] Value of persisted item
+    # @return [nil]
     def put(object_id, value)
       Aspera.assert_type(value, String)
       persist_filepath = id_to_filepath(object_id)
@@ -42,8 +47,11 @@ module Aspera
       File.write(persist_filepath, value)
       Environment.restrict_file_access(persist_filepath)
       @cache[object_id] = value
+      nil
     end
 
+    # Delete persisted item
+    # @param object_id [String] Identifier of persisted item
     def delete(object_id)
       persist_filepath = id_to_filepath(object_id)
       Log.log.debug{"persistency deleting: #{persist_filepath}"}
@@ -51,8 +59,9 @@ module Aspera
       @cache.delete(object_id)
     end
 
+    # Delete persisted items
     def garbage_collect(persist_category, max_age_seconds=nil)
-      garbage_files = Dir[File.join(@folder, persist_category + '*' + FILE_SUFFIX)]
+      garbage_files = current_files(persist_category)
       if !max_age_seconds.nil?
         current_time = Time.now
         garbage_files.select! { |filepath| (current_time - File.stat(filepath).mtime).to_i > max_age_seconds}
@@ -61,7 +70,16 @@ module Aspera
         File.delete(filepath)
         Log.log.debug{"persistency deleted expired: #{filepath}"}
       end
+      @cache.clear
       return garbage_files
+    end
+
+    def current_files(persist_category)
+      Dir[File.join(@folder, persist_category + '*' + FILE_SUFFIX)]
+    end
+
+    def current_items(persist_category)
+      current_files(persist_category).each_with_object({}) {|i, h| h[File.basename(i, FILE_SUFFIX)] = File.read(i)}
     end
 
     private