+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * New Features:
     * `faspex5`: New command: `admin contact reset_password`
+    * `config`: New command: `tokens` with `list`, `show`, `flush` (replace `flush_tokens`)
   * Issues Fixed:
     * `config`: Soft links in transfer SDK archive are correctly extracted
     * `aoc`: Fix `packages delete` not working.